# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API EliteRT",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pembayaran-iuran": {
            "get": {
                "tags": [
                    "Pembayaran Iuran"
                ],
                "summary": "Menampilkan semua data pembayaran iuran",
                "operationId": "b9e931e182fefedaeb8f31ea7d4a9289",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pembayaran Iuran"
                ],
                "summary": "Menambahkan data pembayaran iuran",
                "operationId": "5132e6f807100c8c186db14b8e7d2d60",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_rumah",
                                    "id_penghuni",
                                    "bulan",
                                    "tahun",
                                    "tagihan",
                                    "status_pembayaran"
                                ],
                                "properties": {
                                    "id_rumah": {
                                        "type": "integer"
                                    },
                                    "id_penghuni": {
                                        "type": "integer"
                                    },
                                    "bulan": {
                                        "type": "string",
                                        "example": "Juli"
                                    },
                                    "tahun": {
                                        "type": "integer",
                                        "example": 2025
                                    },
                                    "tagihan": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50000
                                    },
                                    "status_pembayaran": {
                                        "type": "string",
                                        "enum": [
                                            "belum bayar",
                                            "lunas"
                                        ]
                                    },
                                    "metode_pembayaran": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data berhasil ditambahkan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
...
        "/api/pembayaran-iuran/rumah/{id}": {
            "get": {
                "tags": [
                    "Pembayaran Iuran"
                ],
                "summary": "Menampilkan penghuni aktif berdasarkan ID rumah",
                "operationId": "9c0a410f8ee4285e460658cfa0946d67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID rumah",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data penghuni"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
        "/api/pembayaran-iuran/{id}": {
            "put": {
                "tags": [
                    "Pembayaran Iuran"
                ],
                "summary": "Mengupdate data pembayaran iuran",
                "operationId": "a34f0c624b17269a4eedf29c0a93d64e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pembayaran iuran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_rumah",
                                    "id_penghuni",
                                    "bulan",
                                    "tahun",
                                    "tagihan",
                                    "status_pembayaran"
                                ],
                                "properties": {
                                    "id_rumah": {
                                        "type": "integer"
                                    },
                                    "id_penghuni": {
                                        "type": "integer"
                                    },
                                    "bulan": {
                                        "type": "string",
                                        "example": "Juli"
                                    },
                                    "tahun": {
                                        "type": "integer",
                                        "example": 2025
                                    },
                                    "tagihan": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 75000
                                    },
                                    "status_pembayaran": {
                                        "type": "string",
                                        "enum": [
                                            "belum bayar",
                                            "lunas"
                                        ]
                                    },
                                    "metode_pembayaran": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil diupdate"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pembayaran Iuran"
                ],
                "summary": "Menghapus (soft delete) data pembayaran iuran",
                "operationId": "03af7052e23b282ab051648b9fab104a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil dihapus"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
...
    },
...
}